requirements.txt

djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.7,<5.0
Django>=4.0,<5.0  # Django框架的版本，可以根据需要调整
gunicorn>=20.1.0  # 用于在生产环境中运行Django应用的 WSGI 服务器
